pyproject.toml
[project]
name = "funrest"
dynamic = ["version"]
authors = [
  {name = "Tim Music"},
  {name = "Massimo Fasanella"},
  {name = "Dominik Rey"},
  {name = "Thilo Kieser"},
]

maintainers = [
{name = "teamFunrest", email = "contact@projectfunrest.com"}
]

requires-python = ">=3.9"
dependencies = [
  "Werkzeug>=3.1",
  "Jinja2>=3.1.2",
  "itsdangerous>=2.2",
  "click>=8.1.3",
  "blinker>=1.9",
  "importlib-metadata>=3.6; python_version < '3.10'",
  "flask>=3.1",
  "flask-login>=0.6",
  "gunicorn>=23.0",
  "mariadb>=1.1",
  "mysql.connector>=2.2",
  "flask_sqlalchemy>=3.1",
  "mysql>=0.0.2",
  "flask_mysqldb>=2.0",
  "flask_cors>=2.8",
]

[build-system]
requires = [ "setuptools>=41", "wheel", "setuptools-git-versioning>=2.0,<3", ]
build-backend = "setuptools.build_meta"
 
[tool.setuptools-git-versioning]
enabled = true